new commit before real changes in the signup,order to intregate new api

+++ app/src/main/java/com/aashutosh/desimall_pro/adapter/OrderListAdapter.kt
@@ -79,7 +79,7 @@ class OrderListAdapter(
                     detail[1],
                     detail[2],
                     "",
-                    detail[3].toInt(),
+                    detail[3].toDouble().toInt(),
                     detail[4].toDouble(),
                     detail[5].toDouble()
                 )
